docs(examples): one file per demo, scoring inline as turns are captured
Folds the maturity check into the main demo and drops the separate monitor
entry point — the whole story now reads top to bottom in one file.

Scoring runs after each captured turn, so the progression from support 0 to
full arm weight, and coverage climbing as the second intent forms, is visible
where the capture happens rather than in a second script.
