+++ examples/configurable-adaptive-ranking-python/tools.py
@@ -5,11 +5,7 @@ The Python mirror of ``examples/configurable-adaptive-ranking-ts/src/tools.ts``.
 
 from __future__ import annotations
 
-import json
-from collections.abc import Callable, Sequence
-from dataclasses import dataclass
-
-from ratel_ai import ExecutableTool, IntentGraph, ToolCatalog, TraceSinkConfig
+from ratel_ai import ExecutableTool, ToolCatalog, TraceSinkConfig
 
 
 async def build_catalog(trace: TraceSinkConfig | None = None) -> ToolCatalog:
@@ -56,6 +52,7 @@ BASELINE_TURNS = [
     {"turn": "is the build broken again", "invoked": "gh_run_list", "ok": True},
     {"turn": "the build broken on main", "invoked": "gh_run_list", "ok": True},
     {"turn": "why is the build broken", "invoked": "docker_build", "ok": False},
+    {"turn": "rotate the signing key", "invoked": "vault_rotate", "ok": True},
 ]
 """What the customer's agent did on its own, before Ratel ranked anything.
 
@@ -69,64 +66,6 @@ def top_ids(catalog: ToolCatalog, query: str) -> list[str]:
     return [hit.tool_id for hit in catalog.search(query, 3)]
 
 
+
 HELD_OUT = ["why is the build broken", "rotate the signing key"]
-"""Queries the graph has never been trained on — the coverage probe."""
-
-SUPPORT_FULL = 3
-"""Ratel's full-confidence threshold (``SUPPORT_FULL`` in core).
-
-Not yet exposed by the SDK — see the README's "rough edges".
-"""
-
-
-@dataclass(frozen=True)
-class Readiness:
-    """What a maturity check reports about a candidate graph."""
-
-    clusters: int
-    mature: int
-    """Clusters whose arm carries FULL weight. Below this the boost is ramped
-    down proportionally, so a graph of support-1 clusters barely moves ranking."""
-    observations: int
-    seeded: int
-    ghosts: list[str]
-    """Edges naming ids the serving catalog no longer defines. These are dropped
-    silently at rank time, so a non-zero count means the graph looks populated
-    but will boost less than it appears to — or nothing at all."""
-    coverage_hits: int
-    coverage_probed: int
-    """Held-out queries that matched some cluster, over the number probed. The
-    only metric measured against queries the graph has NOT seen: cluster count
-    and support both rise whether or not it generalises."""
-
-
-async def readiness(
-    graph: IntentGraph,
-    serving: ToolCatalog,
-    held_out: Sequence[str] = tuple(HELD_OUT),
-    known: Callable[[str], bool] | None = None,
-) -> Readiness:
-    """Score a candidate graph without attaching it to anything you serve.
-
-    ``serving`` supplies the tool ids to check edges against (override with
-    ``known`` to simulate drift); the coverage probe runs on a throwaway catalog
-    so the graph under test never touches live ranking.
-    """
-    intents = json.loads(graph.to_json())["intents"]
-    defines = known or serving.has
-
-    probe = await build_catalog(TraceSinkConfig(kind="memory", session_id="readiness-probe"))
-    probe.experimental_enable_adaptive_ranking(graph)
-    for query in held_out:
-        probe.search(query, 5)
-    boosts = [e for e in probe.drain_trace_events() if e["type"] == "usage_boost"]
-
-    return Readiness(
-        clusters=graph.cluster_count,
-        mature=sum(1 for it in intents if it["support"] >= SUPPORT_FULL),
-        observations=sum(it["support"] for it in intents),
-        seeded=sum(it.get("seeded_support", 0) for it in intents),
-        ghosts=[tool for it in intents for tool in it["tools"] if not defines(tool)],
-        coverage_hits=sum(1 for e in boosts if e["intent"] is not None),
-        coverage_probed=len(boosts),
-    )
+"""Queries the graph is scored against — deliberately not the ones it trained on."""
