+++ examples/configurable-adaptive-ranking-ts/package.json
@@ -5,7 +5,6 @@
   "type": "module",
   "scripts": {
     "start": "pnpm -F @ratel-ai/sdk build && tsx src/index.ts",
-    "start:monitor": "pnpm -F @ratel-ai/sdk build && tsx src/monitor.ts",
     "typecheck": "tsc"
   },
   "dependencies": {
